Proj/apps/main/classes.py
from rest_framework.pagination import PageNumberPagination
from rest_framework import response


class CustomPageNumberPagination(PageNumberPagination):
    page_size_query_param = 'size'
    page_query_param = 'page'
    max_page_size = 10000

    def get_paginated_response(self, data):
        return response.Response({
            'next': self.get_next_link(),
            'previous': self.get_previous_link(),
            'count': self.page.paginator.count,
            'page_size': self.page_size,
            'result': data
        })
