requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.9
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
clamd==1.0.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==0.6.0.post1
convert-numbers==0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==21.1
django-jazzmin==2.4.8
django-safe-filefield==0.3.1
django-silk==4.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangorestframework-xml==2.0.0
dnspython==2.1.0
drf-yasg==1.20.0
filelock==3.0.12
filetype==1.0.8
gprof2dot==2021.2.21
gunicorn==20.1.0
html5lib==1.1
idna==2.10
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.2.2
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==5.2.0
numpy==1.21.4
openapi-codec==1.3.2
packaging==21.0
pdfkit==1.0.0
persiantools==2.1.1
Pillow==8.1.0
pkg_resources==0.0.0
platformdirs==2.0.2
progress==1.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.7.0
pycparser==2.20
pycurl==7.44.1
Pygments==2.10.0
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-decouple==3.4
python-magic==0.4.22
pytz==2020.5
reportlab==3.6.1
requests==2.25.1
resolvelib==0.5.5
rsa==4.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wkhtmltopdf==0.2
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.7.0
zipp==3.6.0
openpyxl==3.0.9